+++ lib/native_liquid_glass.dart
@@ -18,6 +18,7 @@ export 'src/liquid_glass_stepper.dart';
 export 'src/liquid_glass_tab_bar.dart';
 export 'src/liquid_glass_toggle.dart';
 export 'src/liquid_glass_toolbar.dart';
+export 'src/shares/liquid_glass_border.dart';
 export 'src/shares/liquid_glass_config.dart';
 export 'src/shares/liquid_glass_icon.dart';
 export 'src/utils/liquid_glass_spring.dart';